requirements.txt
flask
Flask-SQLAlchemy
mysql-connector-python
mysqlclient
python-dotenv
flask-restful
flask-apispec
apispec
marshmallow